refactor: move authenticationpage to file
- remove templated comments and code
